+++ src/syscalls.h
@@ -1,5 +1,5 @@
 // syscalls.h
-// v0.14.2 (c) Alexander 'xaitax' Hagenah
+// v0.15.0 (c) Alexander 'xaitax' Hagenah
 // Licensed under the MIT License. See LICENSE file in the project root for full license information.
 
 #ifndef SYSCALLS_H
@@ -9,6 +9,26 @@
 
 #ifndef NTSTATUS
 using NTSTATUS = LONG;
+#endif
+
+#ifndef STATUS_BUFFER_TOO_SMALL
+#define STATUS_BUFFER_TOO_SMALL ((NTSTATUS)0xC0000023L)
+#endif
+
+#ifndef STATUS_BUFFER_OVERFLOW
+#define STATUS_BUFFER_OVERFLOW ((NTSTATUS)0x80000005L)
+#endif
+
+#ifndef OBJ_CASE_INSENSITIVE
+#define OBJ_CASE_INSENSITIVE 0x00000040L
+#endif
+
+#ifndef REG_SZ
+#define REG_SZ 1
+#endif
+
+#ifndef REG_EXPAND_SZ
+#define REG_EXPAND_SZ 2
 #endif
 
 struct SYSCALL_ENTRY
@@ -35,6 +55,10 @@ struct SYSCALL_STUBS
     SYSCALL_ENTRY NtSetContextThread;
     SYSCALL_ENTRY NtResumeThread;
     SYSCALL_ENTRY NtFlushInstructionCache;
+    SYSCALL_ENTRY NtClose;
+    SYSCALL_ENTRY NtOpenKey;
+    SYSCALL_ENTRY NtQueryValueKey;
+    SYSCALL_ENTRY NtEnumerateKey;
 };
 
 struct UNICODE_STRING_SYSCALLS
@@ -110,6 +134,36 @@ struct CLIENT_ID
 };
 using PCLIENT_ID = CLIENT_ID *;
 
+enum KEY_VALUE_INFORMATION_CLASS
+{
+    KeyValueBasicInformation = 0,
+    KeyValueFullInformation,
+    KeyValuePartialInformation
+};
+
+struct KEY_VALUE_PARTIAL_INFORMATION
+{
+    ULONG TitleIndex;
+    ULONG Type;
+    ULONG DataLength;
+    UCHAR Data[1];
+};
+using PKEY_VALUE_PARTIAL_INFORMATION = KEY_VALUE_PARTIAL_INFORMATION *;
+
+struct KEY_BASIC_INFORMATION
+{
+    LARGE_INTEGER LastWriteTime;
+    ULONG TitleIndex;
+    ULONG NameLength;
+    WCHAR Name[1];
+};
+using PKEY_BASIC_INFORMATION = KEY_BASIC_INFORMATION *;
+
+enum KEY_INFORMATION_CLASS
+{
+    KeyBasicInformation = 0
+};
+
 inline void InitializeObjectAttributes(POBJECT_ATTRIBUTES p, PUNICODE_STRING_SYSCALLS n, ULONG a, HANDLE r, PVOID s)
 {
     p->Length = sizeof(OBJECT_ATTRIBUTES);
@@ -119,6 +173,22 @@ inline void InitializeObjectAttributes(POBJECT_ATTRIBUTES p, PUNICODE_STRING_SYS
     p->SecurityDescriptor = s;
     p->SecurityQualityOfService = nullptr;
 }
+
+#ifndef KEY_QUERY_VALUE
+#define KEY_QUERY_VALUE (0x0001)
+#endif
+
+#ifndef KEY_READ
+#define KEY_READ (0x20019)
+#endif
+
+#ifndef KEY_WOW64_64KEY
+#define KEY_WOW64_64KEY (0x0100)
+#endif
+
+#ifndef KEY_WOW64_32KEY
+#define KEY_WOW64_32KEY (0x0200)
+#endif
 
 extern "C"
 {
@@ -141,6 +211,10 @@ extern "C"
     NTSTATUS NtSetContextThread_syscall(HANDLE, PCONTEXT);
     NTSTATUS NtResumeThread_syscall(HANDLE, PULONG);
     NTSTATUS NtFlushInstructionCache_syscall(HANDLE, PVOID, ULONG);
+    NTSTATUS NtClose_syscall(HANDLE);
+    NTSTATUS NtOpenKey_syscall(PHANDLE, ACCESS_MASK, POBJECT_ATTRIBUTES);
+    NTSTATUS NtQueryValueKey_syscall(HANDLE, PUNICODE_STRING_SYSCALLS, KEY_VALUE_INFORMATION_CLASS, PVOID, ULONG, PULONG);
+    NTSTATUS NtEnumerateKey_syscall(HANDLE, ULONG, KEY_INFORMATION_CLASS, PVOID, ULONG, PULONG);
 }
 
 #endif